+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+---
+### `v1.0.0 - 2025.05.19`
+
+Created the Cif class and the CifValidator class.
+Created the cif() helper function as a shortcut to the Cif::from() method.